+++ src/main/java/me/pajic/rearm/ReArm.java
@@ -27,7 +27,7 @@ import me.pajic.rearm.platform.fabric.FabricPlatform;
 public class ReArm {
 
 	public static final String MOD_ID = /*$ mod_id*/ "rearm";
-	public static final String MOD_VERSION = /*$ mod_version*/ "2.4.13";
+	public static final String MOD_VERSION = /*$ mod_version*/ "2.4.15";
 	public static final String MOD_FRIENDLY_NAME = /*$ mod_name*/ "ReArm";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 	public static final Identifier CONFIG_RL = Identifier.fromNamespaceAndPath(MOD_ID, "config");