+++ src/main/java/me/pajic/rearm/platform/fabric/EarlyRiser.java
@@ -17,7 +17,7 @@ public class EarlyRiser implements Runnable {
     private static final String targetType = // Gui$HeartType
             mappingResolver.mapClassName("intermediary", "net.minecraft.class_329$class_6411");
 
-    private static final String paramType = // ResourceLocation / Identifier
+    private static final String paramType = // Identifier
             "L" + mappingResolver.mapClassName("intermediary", "net.minecraft.class_2960") + ";";
 
     @Override