+++ src/main/java/me/pajic/rearm/renderer/LegacyThrownAxeRenderer.java
@@ -11,7 +11,7 @@ import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.renderer.texture.TextureAtlas;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemDisplayContext;
@@ -46,7 +46,7 @@ public class LegacyThrownAxeRenderer extends EntityRenderer<ThrownAxe> {
 
     @SuppressWarnings("deprecation")
     @Override
-    public @NotNull ResourceLocation getTextureLocation(@NotNull ThrownAxe entity) {
+    public @NotNull Identifier getTextureLocation(@NotNull ThrownAxe entity) {
         return TextureAtlas.LOCATION_BLOCKS;
     }
 }